requirements.txt
grpcio == 1.46.0
grpcio-tools == 1.46.0
pytest == 7.1.2
psycopg == 3.0.12
psycopg-binary == 3.0.12
pyYAML==6.0
psutil~=5.9.0
PyJWT~=2.4.0